Module Tjr_mem_queue__With_lwt

Lwt version; since lwt mutexes and cvars can be created outside the monad, we can create the queues outside the monad

module Internal : sig ... end

val make_as_object : unit -> ('msg, Tjr_monad.With_lwt.lwt) Tjr_mem_queue.Memq_intf.memq_as_obj